Update semantics and move /500 to /50x as it is a more general error page
