+++ site/routes/index.js
@@ -4,7 +4,7 @@ import { routeDefinitions } from './definitions';
 import L from '../layout';
 import HomePage from '../pages/home';
 import NotFoundPage from '../pages/not-found';
-import InternalErrorPage from '../pages/internal-error';
+import ServerErrorPage from '../pages/server-error';
 import WhatWeDoPage from '../pages/what-we-do';
 
 const TITLE_SUFFIX = 'Red Badger';
@@ -39,7 +39,7 @@ export default function routes({ contactUsURL }) {
     homePage: () => <L><HomePage contactUsURL={contactUsURL} /></L>,
     notFoundPage: () => <L><NotFoundPage /></L>,
     whatWeDoPage: () => <L><WhatWeDoPage /></L>,
-    internalErrorPage: () => <L><InternalErrorPage /></L>,
+    serverErrorPage: () => <L><ServerErrorPage /></L>,
   };
 
   return prefixRoutes(routeDefinitions.map(
